Refactor CardType to registry annotations
