+++ farebot-app/src/main/java/com/codebutler/farebot/app/core/sample/SampleCard.kt
@@ -27,7 +27,7 @@ import com.codebutler.farebot.base.ui.uiTree
 import com.codebutler.farebot.base.ui.FareBotUiTree
 import com.codebutler.farebot.base.util.ByteArray
 import com.codebutler.farebot.card.Card
-import com.codebutler.farebot.card.CardType
+import com.codebutler.farebot.transit.registry.annotations.CardType
 import java.util.Date
 
 class SampleCard(private val rawCard: RawSampleCard) : Card() {
