+++ farebot-app/src/main/java/com/codebutler/farebot/app/core/serialize/gson/CardKeysGsonTypeAdapterFactory.kt
@@ -22,7 +22,7 @@
 
 package com.codebutler.farebot.app.core.serialize.gson
 
-import com.codebutler.farebot.card.CardType
+import com.codebutler.farebot.transit.registry.annotations.CardType
 import com.codebutler.farebot.card.classic.key.ClassicCardKeys
 import com.codebutler.farebot.key.CardKeys
 import com.google.gson.Gson
