+++ farebot-card-classic/src/main/java/com/codebutler/farebot/card/classic/key/ClassicCardKeys.java
@@ -26,7 +26,7 @@ package com.codebutler.farebot.card.classic.key;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
-import com.codebutler.farebot.card.CardType;
+import com.codebutler.farebot.transit.registry.annotations.CardType;
 import com.codebutler.farebot.key.CardKeys;
 import com.google.auto.value.AutoValue;
 import com.google.gson.Gson;
