+++ farebot-card-classic/src/main/java/com/codebutler/farebot/card/classic/raw/RawClassicCard.java
@@ -26,7 +26,7 @@ package com.codebutler.farebot.card.classic.raw;
 import android.support.annotation.NonNull;
 
 import com.codebutler.farebot.base.util.ByteArray;
-import com.codebutler.farebot.card.CardType;
+import com.codebutler.farebot.transit.registry.annotations.CardType;
 import com.codebutler.farebot.card.RawCard;
 import com.codebutler.farebot.card.classic.ClassicCard;
 import com.codebutler.farebot.card.classic.ClassicSector;
