+++ farebot-card-desfire/src/main/java/com/codebutler/farebot/card/desfire/raw/RawDesfireCard.java
@@ -25,7 +25,7 @@ package com.codebutler.farebot.card.desfire.raw;
 import android.support.annotation.NonNull;
 
 import com.codebutler.farebot.base.util.ByteArray;
-import com.codebutler.farebot.card.CardType;
+import com.codebutler.farebot.transit.registry.annotations.CardType;
 import com.codebutler.farebot.card.RawCard;
 import com.codebutler.farebot.card.desfire.DesfireApplication;
 import com.codebutler.farebot.card.desfire.DesfireCard;
