+++ farebot-card-felica/src/main/java/com/codebutler/farebot/card/felica/FelicaCard.java
@@ -29,7 +29,7 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
 import com.codebutler.farebot.card.Card;
-import com.codebutler.farebot.card.CardType;
+import com.codebutler.farebot.transit.registry.annotations.CardType;
 import com.codebutler.farebot.base.ui.FareBotUiTree;
 import com.codebutler.farebot.base.util.ByteArray;
 import com.google.auto.value.AutoValue;
