+++ farebot-card-felica/src/main/java/com/codebutler/farebot/card/felica/raw/RawFelicaCard.java
@@ -26,7 +26,7 @@ import android.support.annotation.NonNull;
 
 import com.codebutler.farebot.base.util.ByteArray;
 import com.codebutler.farebot.card.Card;
-import com.codebutler.farebot.card.CardType;
+import com.codebutler.farebot.transit.registry.annotations.CardType;
 import com.codebutler.farebot.card.RawCard;
 import com.codebutler.farebot.card.felica.FelicaCard;
 import com.codebutler.farebot.card.felica.FelicaSystem;
