+++ farebot-card-ultralight/src/main/java/com/codebutler/farebot/card/ultralight/UltralightCard.java
@@ -27,7 +27,7 @@ import android.content.Context;
 import android.support.annotation.NonNull;
 
 import com.codebutler.farebot.card.Card;
-import com.codebutler.farebot.card.CardType;
+import com.codebutler.farebot.transit.registry.annotations.CardType;
 import com.codebutler.farebot.base.ui.FareBotUiTree;
 import com.codebutler.farebot.base.util.ByteArray;
 import com.google.auto.value.AutoValue;
