+++ farebot-card/build.gradle
@@ -2,6 +2,7 @@ apply plugin: 'com.android.library'
 
 dependencies {
     api project(':farebot-base')
+    api project(':farebot-transit-registry-annotations')
 
     api libs.supportV4
     api libs.gson
